Unify template match method into a single enum across domain and match.
